pyproject.toml
[tool.poetry]
name = "jua-zarr-api"
version = "0.1.0"
description = ""
authors = ["Azamat <azamat@galimzhanov.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.10"
aws-cdk-lib = "^2.60.0"
fastapi = "^0.89.1"
mangum = "^0.17.0"
xarray = "^2022.12.0"
zarr = "^2.13.3"
s3fs = "^2022.11.0"
fsspec = "^2022.11.0"
fastapi-utils = "^0.2.1"
scipy = "^1.10.0"
sentry-sdk = {extras = ["fastapi"], version = "^1.13.0"}
bottleneck = "^1.3.6"
metpy = "^1.4.0"
aiobotocore = {extras = ["boto3"], version = "^2.4.2"}
mypy = "^1.0.0"
boto3-stubs = "^1.26.68"
types-python-dateutil = "^2.8.19.6"

[tool.poetry.group.test.dependencies]
pytest = "^7.2.0"
pytest-cov = "^4.0.0"
httpx = "^0.23.3"


[tool.poetry.group.format.dependencies]
black = "^22.10.0"
isort = "^5.10.1"
pylint = "^2.15.5"
pre-commit = "^2.20.0"
commitizen = "^2.37.0"


[tool.poetry.group.dev.dependencies]
uvicorn = {extras = ["standard"], version = "^0.20.0"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"



[tool.poetry-exec-plugin.commands]
test = "pytest"
test-watch = "find . -name '*.py' | entr poetry exec test"

format = "bash ./scripts/format.sh"
format-ci = "bash ./scripts/format-ci.sh"

serve = "PYTHONPATH=. APP_ROOT_PATH=/dev python app/main.py"
proxy = "traefik --configFile=traefik.toml"

deploy = "cdk deploy"
deploy-ci = "cdk deploy --require-approval never"
synth = "cdk synth"
teardown-ci = "cdk destroy -f"


[tool.isort]
profile = 'black'

[tool.pylint.format]
max-line-length = "88"
max-args = "7"

[tool.pylint.MASTER]
ignore ="cdk.out,.venv"
disable = "E0611,"

[tool.pylint."MESSAGES CONTROL"]
enable = "useless-suppression"
disable = [
    "too-few-public-methods",
]

[tool.pytest.ini_options]
addopts = "--cov-report=html --cov-report=term-missing --cov=app --ignore=cdk.out"

[tool.mypy]
warn_unused_configs = true
exclude = [
    "cdk.out",
]

[[tool.mypy.overrides]]
module = [
    "metpy.*",
    "pandas",
]
ignore_missing_imports = true
